tests/requirements.txt
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
PyHamcrest==1.9.0
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
